Refactor word method
